fix: Member 관련 DTO 파일 위치 변경
