+++ src/main/java/chathealth/chathealth/controller/IndexController.java
@@ -1,11 +1,10 @@
 package chathealth.chathealth.controller;
 
 import chathealth.chathealth.constants.Role;
-import chathealth.chathealth.dto.response.CustomUserDetails;
+import chathealth.chathealth.dto.response.member.CustomUserDetails;
 import chathealth.chathealth.entity.member.Member;
 import chathealth.chathealth.service.MemberService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
