+++ src/main/java/chathealth/chathealth/dto/request/member/UserEditDto.java
@@ -1,4 +1,4 @@
-package chathealth.chathealth.dto.request;
+package chathealth.chathealth.dto.request.member;
 
 import chathealth.chathealth.entity.member.Address;
 import jakarta.validation.constraints.NotBlank;
