+++ src/main/java/chathealth/chathealth/dto/request/member/UserJoinDto.java
@@ -1,6 +1,7 @@
-package chathealth.chathealth.dto.request;
+package chathealth.chathealth.dto.request.member;
 
 import chathealth.chathealth.entity.member.Address;
+import chathealth.chathealth.constants.Grade;
 import chathealth.chathealth.constants.Role;
 import jakarta.validation.constraints.Pattern;
 import lombok.Builder;
@@ -16,7 +17,7 @@ import java.time.LocalDateTime;
 @Setter
 @Builder
 @ToString
-public class EntJoinDto {
+public class UserJoinDto {
     private Long id;
 
     @Pattern(regexp = "^(?=.*[a-zA-Z])(?=.*[!@#$%^*+=-])(?=.*[0-9]).{8,15}$",message = "비밀번호는 8~15자리이고 영문, 숫자, 기호를 포함하고 있어야 합니다.")
@@ -26,12 +27,11 @@ public class EntJoinDto {
     private String postcode;
     private String frontAddress;
     private String addressDetail;
-    private String company;
-    private String ceo;
-    private String entNo;
+    private String name;
+    private String nickname;
+    private LocalDate birth;
     private MultipartFile profile;
     private Role role;
-    private LocalDateTime createdDate;
-    private LocalDate birth;
+    private Grade grade;
+    private LocalDateTime createDate;
 }
-
