+++ src/main/java/chathealth/chathealth/dto/response/member/EntInfoDto.java
@@ -1,4 +1,4 @@
-package chathealth.chathealth.dto.response;
+package chathealth.chathealth.dto.response.member;
 
 import chathealth.chathealth.entity.member.Address;
 import chathealth.chathealth.constants.Role;
