+++ src/main/java/chathealth/chathealth/entity/member/Ent.java
@@ -1,6 +1,6 @@
 package chathealth.chathealth.entity.member;
 
-import chathealth.chathealth.dto.request.EntEditDto;
+import chathealth.chathealth.dto.request.member.EntEditDto;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import lombok.Getter;
