+++ src/main/java/chathealth/chathealth/entity/member/Users.java
@@ -1,7 +1,7 @@
 package chathealth.chathealth.entity.member;
 
 import chathealth.chathealth.constants.Grade;
-import chathealth.chathealth.dto.request.UserEditDto;
+import chathealth.chathealth.dto.request.member.UserEditDto;
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
@@ -9,7 +9,6 @@ import jakarta.persistence.Enumerated;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.SuperBuilder;
-import org.hibernate.annotations.DynamicInsert;
 
 import static lombok.AccessLevel.PROTECTED;
