+++ src/main/java/chathealth/chathealth/repository/ReViewRepository.java
@@ -1,14 +1,11 @@
 package chathealth.chathealth.repository;
 
-import chathealth.chathealth.dto.request.BoardEditDto;
-import chathealth.chathealth.dto.request.ReviewModDto;
 import chathealth.chathealth.entity.Review;
 import chathealth.chathealth.entity.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface ReViewRepository extends JpaRepository<Review,Long> {
 
@@ -16,5 +13,6 @@ public interface ReViewRepository extends JpaRepository<Review,Long> {
     List<Review> findAllByPost(Post post);
 
     List<Review> findAllById(long id);
+    List<Review> findByMember(long id);
 
 }
