+++ src/main/java/chathealth/chathealth/repository/boardComment/BoardCommentRepositoryImpl.java
@@ -1,7 +1,7 @@
 package chathealth.chathealth.repository.boardComment;
 
 import chathealth.chathealth.dto.response.BoardCommentResponse;
-import chathealth.chathealth.dto.response.CustomUserDetails;
+import chathealth.chathealth.dto.response.member.CustomUserDetails;
 import chathealth.chathealth.entity.BoardComment;
 import chathealth.chathealth.entity.member.Users;
 import com.querydsl.jpa.impl.JPAQueryFactory;
